requirements/prod.txt
# Everything needed in production

# Flask
click>=7.0
Flask==2.2.2

# Database
Flask-SQLAlchemy==3.0.2
psycopg2-binary==2.9.4
SQLAlchemy==1.4.41

# Migrations
Flask-Migrate==3.1.0

# Deployment
gunicorn>=19.9.0

# Auth
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.4.4 

# Environment variable parsing
environs==9.5.0

# Serialization/deserialization
marshmallow-sqlalchemy==0.28.1
flask-marshmallow==0.14.0

# Error Tracking
sentry-sdk[flask]==1.9.10
